+++ src/birth_death_generation_method.c
@@ -27,11 +27,11 @@ static RET_VAL _DoTransformation( ABSTRACTION_METHOD *method, IR *ir, SPECIES *s
 
 static RET_VAL _CombineDegradationReactions(ABSTRACTION_METHOD *method, IR *ir, SPECIES *species );
 static RET_VAL _CombineProductionReactions( ABSTRACTION_METHOD *method, IR *ir, SPECIES *species );
-static RET_VAL _AdjustProductionKineticLaw( ABSTRACTION_METHOD *method, SPECIES *species, REACTION *reaction, KINETIC_LAW *productionKineticLaw, KINETIC_LAW *degradationKineticLaw, long stoichiometry );
-static RET_VAL _AdjustDegradationKineticLaw( ABSTRACTION_METHOD *method, SPECIES *species, REACTION *reaction, KINETIC_LAW *productionKineticLaw, KINETIC_LAW *degradationKineticLaw, long stoichiometry );
+static RET_VAL _AdjustProductionKineticLaw( ABSTRACTION_METHOD *method, SPECIES *species, REACTION *reaction, KINETIC_LAW *productionKineticLaw, KINETIC_LAW *degradationKineticLaw, double stoichiometry );
+static RET_VAL _AdjustDegradationKineticLaw( ABSTRACTION_METHOD *method, SPECIES *species, REACTION *reaction, KINETIC_LAW *productionKineticLaw, KINETIC_LAW *degradationKineticLaw, double stoichiometry );
 static RET_VAL _AddModifiers( ABSTRACTION_METHOD *method, IR *ir, REACTION *production, REACTION *degradation );
-static KINETIC_LAW *_GenerateReplacementForProduction( ABSTRACTION_METHOD *method, SPECIES *species, long stoichiometry ); 
-static KINETIC_LAW *_GenerateReplacementForDegradation( ABSTRACTION_METHOD *method, SPECIES *species, long stoichiometry ); 
+static KINETIC_LAW *_GenerateReplacementForProduction( ABSTRACTION_METHOD *method, SPECIES *species, double stoichiometry ); 
+static KINETIC_LAW *_GenerateReplacementForDegradation( ABSTRACTION_METHOD *method, SPECIES *species, double stoichiometry ); 
 
 
 ABSTRACTION_METHOD *BirthDeathGenerationMethodConstructor(  ABSTRACTION_METHOD_MANAGER *manager ) {
@@ -83,7 +83,7 @@ static RET_VAL _ApplyBirthDeathGenerationMethod( ABSTRACTION_METHOD *method, IR 
 }      
 
 static BOOL _IsConditionSatisfied( ABSTRACTION_METHOD *method, SPECIES *species ) {
-    long stoichiometry = -1;
+    double stoichiometry = -1;
     SPECIES *reactant = NULL;
     SPECIES *product = NULL;
     IR_EDGE *edge = NULL;
@@ -160,7 +160,7 @@ static BOOL _IsConditionSatisfied( ABSTRACTION_METHOD *method, SPECIES *species 
 
 static RET_VAL _DoTransformation( ABSTRACTION_METHOD *method, IR *ir, SPECIES *species ) {
     RET_VAL ret = SUCCESS;
-    long stoichiometry = 0L;
+    double stoichiometry = 0L;
     IR_EDGE *productionEdge = NULL;
     IR_EDGE *degradationEdge = NULL;    
     REACTION *production = NULL;
@@ -209,7 +209,7 @@ static RET_VAL _DoTransformation( ABSTRACTION_METHOD *method, IR *ir, SPECIES *s
 
 static RET_VAL _CombineDegradationReactions(ABSTRACTION_METHOD *method, IR *ir, SPECIES *species ) {
     RET_VAL ret = SUCCESS;
-    long stoichiometry = 0;
+    double stoichiometry = 0;
     REACTION *newReaction = NULL;
     KINETIC_LAW *newKineticLaw = NULL;
     REACTION *reaction = NULL;
@@ -258,7 +258,7 @@ static RET_VAL _CombineDegradationReactions(ABSTRACTION_METHOD *method, IR *ir, 
 
 static RET_VAL _CombineProductionReactions(ABSTRACTION_METHOD *method, IR *ir, SPECIES *species ) {
     RET_VAL ret = SUCCESS;
-    long stoichiometry = 0;
+    double stoichiometry = 0;
     REACTION *newReaction = NULL;
     KINETIC_LAW *newKineticLaw = NULL;
     REACTION *reaction = NULL;
@@ -309,7 +309,7 @@ static RET_VAL _CombineProductionReactions(ABSTRACTION_METHOD *method, IR *ir, S
 
 static RET_VAL _AdjustProductionKineticLaw( 
 ABSTRACTION_METHOD *method, SPECIES *species, REACTION *reaction, KINETIC_LAW *productionKineticLaw, 
-KINETIC_LAW *degradationKineticLaw, long stoichiometry ) {
+KINETIC_LAW *degradationKineticLaw, double stoichiometry ) {
     RET_VAL ret = SUCCESS;
     KINETIC_LAW *temp1 = NULL;
     KINETIC_LAW *temp2 = NULL;
@@ -322,7 +322,7 @@ KINETIC_LAW *degradationKineticLaw, long stoichiometry ) {
     
     temp2 = CreateOpKineticLaw( KINETIC_LAW_OP_PLUS, CloneKineticLaw( productionKineticLaw ), CloneKineticLaw( degradationKineticLaw ) );
     temp2 = CreateOpKineticLaw( KINETIC_LAW_OP_DIVIDE, CloneKineticLaw( degradationKineticLaw ), temp2 );
-    temp2 = CreateOpKineticLaw( KINETIC_LAW_OP_POW, temp2, CreateRealValueKineticLaw( (stoichiometry>>1) + 1.0 ) );
+    temp2 = CreateOpKineticLaw( KINETIC_LAW_OP_POW, temp2, CreateRealValueKineticLaw( ((long)stoichiometry>>1) + 1.0 ) );
     temp2 = CreateOpKineticLaw( KINETIC_LAW_OP_MINUS, CreateRealValueKineticLaw( 1.0 ), temp2 );         
     temp3 = CreateOpKineticLaw( KINETIC_LAW_OP_PLUS, CloneKineticLaw( productionKineticLaw ), CloneKineticLaw( degradationKineticLaw ) ); 
     temp2 = CreateOpKineticLaw( KINETIC_LAW_OP_TIMES, temp3, temp2 );    
@@ -340,7 +340,7 @@ KINETIC_LAW *degradationKineticLaw, long stoichiometry ) {
     return SUCCESS;
 }
 
-static RET_VAL _AdjustDegradationKineticLaw( ABSTRACTION_METHOD *method, SPECIES *species, REACTION *reaction, KINETIC_LAW *productionKineticLaw, KINETIC_LAW *degradationKineticLaw, long stoichiometry ) {
+static RET_VAL _AdjustDegradationKineticLaw( ABSTRACTION_METHOD *method, SPECIES *species, REACTION *reaction, KINETIC_LAW *productionKineticLaw, KINETIC_LAW *degradationKineticLaw, double stoichiometry ) {
     RET_VAL ret = SUCCESS;
     KINETIC_LAW *temp1 = NULL;
     KINETIC_LAW *temp2 = NULL;
@@ -349,7 +349,7 @@ static RET_VAL _AdjustDegradationKineticLaw( ABSTRACTION_METHOD *method, SPECIES
     
     temp1 = CreateOpKineticLaw( KINETIC_LAW_OP_DIVIDE, CloneKineticLaw( degradationKineticLaw ), CloneKineticLaw( productionKineticLaw ) );
     temp1 = CreateOpKineticLaw( KINETIC_LAW_OP_PLUS, CreateRealValueKineticLaw( 1.0 ), temp1 );
-    temp1 = CreateOpKineticLaw( KINETIC_LAW_OP_POW, temp1, CreateRealValueKineticLaw( (stoichiometry>>1) + 1.0 ) );
+    temp1 = CreateOpKineticLaw( KINETIC_LAW_OP_POW, temp1, CreateRealValueKineticLaw( ((long)stoichiometry>>1) + 1.0 ) );
     temp1 = CreateOpKineticLaw( KINETIC_LAW_OP_MINUS, temp1, CreateRealValueKineticLaw( 1.0 ) );
     newKineticLaw = CreateOpKineticLaw( KINETIC_LAW_OP_DIVIDE, CloneKineticLaw( degradationKineticLaw ), temp1 );
     
@@ -405,24 +405,24 @@ static RET_VAL _AddModifiers( ABSTRACTION_METHOD *method, IR *ir, REACTION *prod
 
 
 static KINETIC_LAW *_GenerateReplacementForProduction( 
-ABSTRACTION_METHOD *method, SPECIES *species, long stoichiometry ) {
+ABSTRACTION_METHOD *method, SPECIES *species, double stoichiometry ) {
     KINETIC_LAW *replacement = NULL;
     
     replacement = CreateSpeciesKineticLaw( species );
     replacement = CreateOpKineticLaw( KINETIC_LAW_OP_PLUS, replacement, 
-        CreateRealValueKineticLaw( (double)(stoichiometry>>1) ) );
+				      CreateRealValueKineticLaw( (double)((long)stoichiometry>>1) ) );
     return replacement;                      
     
 }
 
 
 static KINETIC_LAW *_GenerateReplacementForDegradation( 
-ABSTRACTION_METHOD *method, SPECIES *species, long stoichiometry ) {
+ABSTRACTION_METHOD *method, SPECIES *species, double stoichiometry ) {
     KINETIC_LAW *replacement = NULL;
     
     replacement = CreateSpeciesKineticLaw( species );
     replacement = CreateOpKineticLaw( KINETIC_LAW_OP_MINUS, replacement, 
-        CreateRealValueKineticLaw( (double)(stoichiometry>>1) ) );
+				      CreateRealValueKineticLaw( (double)((long)stoichiometry>>1) ) );
     return replacement;                          
 }
